### 2022-04-27 #4355 @editorialbot
[PRE REVIEW]: NiaARM: A minimalistic framework for Numerical Association Rule Mining
**Submitting author:** @zStupan (<a href="http://orcid.org/0000-0001-9847-7306">Žiga Stupan</a>)
**Repository:** https://github.com/firefly-cpp/NiaARM
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.6
