+++ pyproject.toml
@@ -10,6 +10,8 @@ include = [
     { path="CHANGELOG.md", format="sdist" },
 ]
 
+exclude = ["niaarm/tests/"]
+
 [tool.poetry.dependencies]
 python = "^3.7"
 niapy = "^2.0.1"
