### 2022-05-23 #4355 @fabian-s
:wave: @juliensiebert @ahurriyetoglu
would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
