### 2022-05-25 a74339b
added text preprocessing funtionalities
