+++ pyproject.toml
@@ -21,6 +21,7 @@ pandas = [
     { version = "^1.3.5", python = ">=3.7.1,<3.8" },
     { version = "^1.4.0", python = "^3.8" }
 ]
+nltk = "^3.7"
 Sphinx = {version = "^4.4.0", optional = true}
 sphinx-rtd-theme = {version = "^1.0.0", optional = true}
 sphinxcontrib-bibtex = {version = "^2.4.1", optional = true}
