### 2022-05-26 dcddbb4
added association rules text mining
