+++ niaarm/text.py
@@ -6,9 +6,21 @@ from collections import Counter
 from pathlib import Path
 from nltk.corpus import stopwords
 from nltk.tokenize import word_tokenize
+from niaarm.rule import Rule
+from niaarm.niaarm import NiaARM, _cut_point
 
 
 class Document:
+    """A text document class.
+
+    Args:
+        text (str): Document text.
+        language (str): Document language. Used for tokenization and stopword removal. Default: 'english'.
+        remove_stopwords (bool): If ``True``, remove stopwords from text. Default: ``True``.
+        lowercase (bool): If ``True``, make text lowercase. Default: ``True``.
+
+    """
+
     def __init__(self, text, language='english', remove_stopwords=True, lowercase=True):
         if lowercase:
             text = text.lower()
@@ -22,10 +34,16 @@ class Document:
 
         self.frequencies = Counter(self.terms)
 
-    def unique(self):
-        return len(self.frequencies)
-
     def frequency(self, term):
+        """Get the frequency of a term,
+
+        Args:
+            term (str): Term to get frequency of.
+
+        Returns:
+            float: Frequency of the term.
+
+        """
         return self.frequencies[term] / len(self)
 
     def __getitem__(self, i):
@@ -39,18 +57,50 @@ class Document:
 
     @classmethod
     def from_file(cls, path, encoding='utf-8', language='english', remove_stopwords=True, lowercase=True):
+        """Construct document from a plain text file.
+
+        Args:
+            path (str): Path to file.
+            encoding (str): Encoding of the file. Default: 'utf-8'.
+            language (str): Language of the file. Default: 'english'.
+            remove_stopwords (bool): If ``True``, remove stopwords from text. Default: ``True``.
+            lowercase (bool): If ``True``, make text lowercase. Default: ``True``.
+
+        Returns:
+            Document: The constructed document.
+
+        """
         text = Path(path).read_text(encoding)
         return cls(text, language, remove_stopwords, lowercase)
 
 
 class Corpus:
+    """The text corpus class.
+
+    Args:
+        documents (Optional[list[Document]]): List of documents. If ``None``, an empty list will be created. Default: ``None``.
+
+    """
+
     def __init__(self, documents=None):
         self.documents = documents or []
 
     def append(self, document):
+        """Add a document to the corpus.
+
+        Args:
+            document (Document): Document to append.
+
+        """
         self.documents.append(document)
 
     def terms(self):
+        """Get a list of unique terms in the corpus
+
+        Returns:
+            list[str]: List of unique terms in the corpus.
+
+        """
         all_terms = set(self.documents[0].terms)
         for doc in self.documents[1:]:
             all_terms.update(doc.terms)
@@ -64,30 +114,175 @@ class Corpus:
 
     @classmethod
     def from_list(cls, lst, language='english', remove_stopwords=True, lowercase=True):
+        """Construct corpus from a list of strings.
+
+        Args:
+            lst (list[str]): List of documents as strings.
+            language (str): Language of the file. Default: 'english'.
+            remove_stopwords (bool): If ``True``, remove stopwords from text. Default: ``True``.
+            lowercase (bool): If ``True``, make text lowercase. Default: ``True``.
+
+        Returns:
+            Corpus: The constructed corpus.
+
+        """
         return cls([Document(text, language, remove_stopwords, lowercase) for text in lst])
 
     @classmethod
     def from_directory(cls, path, encoding='utf-8', language='english', remove_stopwords=True, lowercase=True):
[…]
