### 2022-05-27 d080e06
added max_terms attribute to NiaARTM
