### 2022-05-28 d30c729
Removed docstring from internal build_rule method
