+++ docs/conf.py
@@ -21,7 +21,7 @@ copyright = '2022, Žiga Stupan, Iztok Fister Jr.'
 author = 'Žiga Stupan, Iztok Fister Jr.'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.5'
+release = '0.2.0'
 
 
 # -- General configuration ---------------------------------------------------
