+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "niaarm"
-version = "0.1.6"
+version = "0.2.0"
 description = "A minimalistic framework for numerical association rule mining"
 authors = ["Žiga Stupan <ziga.stupan1@student.um.si>", "Iztok Fister Jr. <iztok.fister1@um.si>"]
 readme = "README.md"
