### 2022-06-01 #4355 @diegomcarvalho
Hi, I’m available. Sorry for the delay… All the best, d.c

Prof. Diego Carvalho, D.Sc

Associate Professor / Production Engineering Department
Graduate Programme in Production Engineering and Systems
Federal Centre for Engineering Studies and Technological Education - CEFET/RJ
Av. Maracanã, 229 - Maracanã
Rio de Janeiro - RJ - 20271-110 - Brasil/Brazil/Brésil
Il sudore non e santo, ma fa miracoli. :wq

what3words pt:dueto.digite.caco en:blazed.nurses.gazed
