### 2022-06-03 a5902ea
Installation instructions for Fedora Linux
