### 2022-07-18 #4448 @fabian-s
@zStupan to speed up this process, you could begin to adress some of the issues identified by @diegomcarvalho while we wait for @timClicks to review the paper.

If you have questions or comments about specific points of the review or what would be required to adress them, please create separate GH issues for them in your repository and link to them here so we can have focussed discussions about them there and not in one sprawling thread.
