+++ README.md
@@ -41,7 +41,7 @@ Install NiaARM with pip:
 pip install niaarm
 ```
 
-To install NiaARM on Alpine Linux, please enable Testing repository and use:
+To install NiaARM on Alpine Linux, please enable Community repository and use:
 
 ```sh
 $ apk add py3-niaarm
