### 2022-09-11 #4448 @fabian-s
@arfon yes, sorry, seems so. I lost track of this over the summer.

@zStupan very sorry for the delay. I tend to agree that the open points left in @diegomcarvalho checklist all seem to be present in the repo.

@diegomcarvalho please confirm that the items missing in your checklist are satisfied in the repo or open issues/provide some additional feedback on the specific points that you feel need additional work by the submitting authors

@timClicks are you still willing to do a review for JOSS for this work?
