### 2022-09-16 #4448 @diegomcarvalho
Hi @fabian-s, all items are satisfied by the collection provided in the repository.

Cheers, d.c
