### 2022-09-28 #4448 @fabian-s
hi @zStupan, i'm filling in as the 2nd reviewer here, sorry again for the delays.

Almost everything seems fine, but please see firefly-cpp/NiaARM#66 for one small request to add a more explicit comparison to other packages to your statement of need section.

I also had to run `import nltk; nltk.download('stopwords'); nltk.download('punkt') ` to get the textmining example to run, but I admittedly did not use `poetry install` as your docs recommend so maybe that was my own fault....
