### 2022-09-28 f23ce53
Update getting_started.rst
