+++ paper/paper.bib
@@ -181,3 +181,14 @@
 	note = {arXiv:2010.10884 [cs]},
 	keywords = {Computer Science - Databases, Computer Science - Neural and Evolutionary Computing},
 }
+
+@article{alcala2009keel,
+  title={KEEL: a software tool to assess evolutionary algorithms for data mining problems},
+  author={Alcal{\'a}-Fdez, Jes{\'u}s and Sanchez, Luciano and Garcia, Salvador and del Jesus, Maria Jose and Ventura, Sebastian and Garrell, Josep Maria and Otero, Jos{\'e} and Romero, Crist{\'o}bal and Bacardit, Jaume and Rivas, Victor M and others},
+  journal={Soft Computing},
+  volume={13},
+  number={3},
+  pages={307--318},
+  year={2009},
+  publisher={Springer}
+}
