### 2022-09-28 firefly-cpp/NiaARM#66 @zStupan
@fabian-s

I've updated the statement of need section as requested. I've also added a note to the docs about having to download stopwords and the punkt tokenizer as you've pointed out.
