### 2022-09-29 #4448 @fabian-s
@zStupan seems like we're finally *almost* there -- please add the missing DOI (see above) and then:

- [ ] Check that your archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
