+++ paper/paper.bib
@@ -184,6 +184,7 @@
 
 @article{alcala2009keel,
   title={KEEL: a software tool to assess evolutionary algorithms for data mining problems},
+  doi = {10.1007/s00500-008-0323-y},
   author={Alcal{\'a}-Fdez, Jes{\'u}s and Sanchez, Luciano and Garcia, Salvador and del Jesus, Maria Jose and Ventura, Sebastian and Garrell, Josep Maria and Otero, Jos{\'e} and Romero, Crist{\'o}bal and Bacardit, Jaume and Rivas, Victor M and others},
   journal={Soft Computing},
   volume={13},
