### 2022-09-29 #4448 @zStupan
@fabian-s the version tag is `0.2.1` and the DOI is `10.5281/zenodo.7123879`
