### 2022-09-29 5fd2110
Small paper fixes
