### 2022-09-29 #4448 @arfon
Acknowledging that @fabian-s provided the second review here.

@zStupan – could you please merge this PR which removes the AAS-related stuff which doesn't apply here.
